pyproject.toml
[project]
name = "otrazhenie"
version = "0.0.1"
description = "MVP платформы для анализа косметики"
authors = [
    {name = "nikewalce",email = "nikewalce1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "flask (==2.3.2)",
    "python-dotenv (==1.0.0)",
    "unidecode (==1.3.6)",
    "pyzbar (>=0.1.9,<0.2.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-login (>=0.6.3,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
